requirements.txt

arrow==1.1.1
asgiref==3.2.10
binaryornot==0.4.4
botocore==1.17.31
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cookiecutter==1.7.3
cryptography==3.4.8
defusedxml==0.7.1
dj-database-url==0.5.0
djangify==1.0.0
Django==3.0.8
django-bootstrap3==15.0.0
django-crispy-forms==1.12.0
django-environ==0.6.0
django-filter==2.4.0
django-js-asset==1.2.2
django-on-heroku==1.1.2
django-phonenumber-field==5.2.0
docutils==0.15.2
environ==1.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
Jinja2==3.0.1
jinja2-time==0.2.0
jmespath==0.10.0
MarkupSafe==2.0.1
mysqlclient==2.0.3
oauth2==1.9.0.post1
oauthlib==3.1.1
phonenumberslite==8.12.33
Pillow==8.3.1
poyo==0.5.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-slugify==5.0.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
socialauth==0.2.0
sqlparse==0.3.1
text-unidecode==1.3
tzdata==2021.1
urllib3==1.25.10
whitenoise==5.1.0
